Move getHeadlessConfigPage to functions
